Migrate to console application (for now WinExe)
